test(generators): move transient-error tests to test_openai_compatible.py
Base-class retry behaviour (transient_retry_codes, GeneratorBackoffTrigger,
terminal-code short-circuit, generic-exception message) belongs in
test_openai_compatible.py.  test_nim.py now contains only NIM-specific
tests that are not inherited from OpenAICompatible.

Refs: #1967
